requirements.txt
loguru>=0.5.3
ruamel.yaml>=0.17.0
requests>=2.25.1
pyfiglet>=0.8.post1